# docs/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Antaride",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/v1/auth/otp/request": {
            "post": {
                "operationId": "v1.auth.otp.request",
                "description": "Response-nya SELALU sama bentuknya, terdaftar maupun belum.",
                "summary": "Minta kode OTP",
                "tags": [
                    "Otp"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestOtpRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "anyOf": [
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "phone_masked": {
                                                            "type": "string",
                                                            "description": "Nomor dikembalikan dalam bentuk tersamarkan, bukan penuh.\nAplikasi sudah tahu nomor yang dia kirim; yang ditampilkan di\nlayar \"masukkan kode\" adalah konfirmasi, dan nomor penuh di sana\nhanya menambah data yang bisa dibaca orang di sekitar."
                                                        },
                                                        "purpose": {
                                                            "type": "string"
                                                        },
                                                        "expires_at": {
                                                            "type": "string"
                                                        },
                                                        "expires_in_seconds": {
                                                            "type": [
                                                                "object",
                                                                "null"
                                                            ]
                                                        },
                                                        "resend_after_seconds": {
                                                            "type": "integer"
                                                        },
                                                        "debug_code": {
                                                            "type": [
                                                                "string",
                                                                "null"
                                                            ]
                                                        }
                                                    },
                                                    "required": [
                                                        "phone_masked",
                                                        "purpose",
                                                        "expires_at",
                                                        "expires_in_seconds",
                                                        "resend_after_seconds",
                                                        "debug_code"
                                                    ]
                                                },
                                                {
                                                    "type": "object",
                                                    "additionalProperties": {
                                                        "type": "string"
                                                    }
                                                },
                                                {
                                                    "type": "string"
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/v1/me": {
            "get": {
                "operationId": "v1.me.show",
                "tags": [
                    "Profile"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "anyOf": [
                                                {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                {
                                                    "type": "object",
                                                    "additionalProperties": {
                                                        "type": "string"
                                                    }
                                                },
                                                {
                                                    "type": "string"
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "v1.me.update",
                "tags": [
                    "Profile"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "anyOf": [
                                                {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                {
                                                    "type": "object",
                                                    "additionalProperties": {
                                                        "type": "string"
                                                    }
                                                },
                                                {
                                                    "type": "string"
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "v1.me.deletion.request",
                "description": "========================================================================\n DITUNDA, BUKAN LANGSUNG DIHAPUS\n========================================================================\n Yang dilakukan hanya menandai `deletion_requested_at`. Penghapusan\n sebenarnya dijalankan job setelah masa tunggu.\n\n Tiga alasan, dan ketiganya nyata:\n\n   1. Penghapusan yang tidak sengaja bisa dibatalkan. Masuk kembali dalam\n      masa tunggu membatalkan pengajuannya.\n   2. Order yang sedang berjalan harus selesai lebih dulu \u2014 termasuk\n      pembayarannya. Menghapus akun di tengah order berarti driver\n      mengantar penumpang yang datanya sudah hilang.\n   3. Ledger TIDAK boleh ikut terhapus. Riwayat keuangan wajib disimpan\n      untuk kewajiban pelaporan, dan `wallet_transactions` bersifat\n      append-only yang ditegakkan trigger database. Yang dihapus adalah\n      data pribadinya, bukan jejak uangnya.\n========================================================================",
                "summary": "Ajukan penghapusan akun",
                "tags": [
                    "Profile"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "anyOf": [
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "deletion_requested_at": {
                                                            "type": "string"
                                                        },
                                                        "scheduled_for": {
                                                            "type": "string"
                                                        },
                                                        "grace_days": {
                                                            "type": "integer"
                                                        },
                                                        "message": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "required": [
                                                        "deletion_requested_at",
                                                        "scheduled_for",
                                                        "grace_days",
                                                        "message"
                                                    ]
                                                },
                                                {
                                                    "type": "object",
                                                    "additionalProperties": {
                                                        "type": "string"
                                                    }
                                                },
                                                {
                                                    "type": "string"
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "error": {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "const": "USER_HAS_ACTIVE_ORDER"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "const": "Selesaikan dulu order yang sedang berjalan sebelum menghapus akun."
                                                },
                                                "details": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "message",
                                                "details"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "error"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RequestOtpRequest": {
                "type": "object",
                "description": "Minta kode OTP.",
                "properties": {
                    "phone": {
                        "type": "string",
                        "description": "Validasi bentuknya SENGAJA longgar di sini. Normalisasi dan pemeriksaan bentuk yang sebenarnya ada di\n`PhoneNumber::normalize()`, dan itu satu-satunya tempat yang tahu\nbentuk kanoniknya. Menuliskan regex nomor HP di sini berarti ada\nDUA definisi \"nomor HP yang valid\" yang harus sepakat \u2014 dan yang\ndi FormRequest akan tertinggal, sehingga nomor yang sah ditolak\nsebelum Action-nya sempat menormalkannya.\n\nYang diperiksa di sini hanya yang tidak mungkin: terlalu pendek\nuntuk nomor apa pun, atau terlalu panjang untuk masuk akal.",
                        "minLength": 9,
                        "maxLength": 20
                    },
                    "purpose": {
                        "type": "string",
                        "enum": [
                            "login",
                            "register",
                            "change_phone"
                        ]
                    }
                },
                "required": [
                    "phone"
                ],
                "title": "RequestOtpRequest"
            },
...
            "UpdateProfileRequest": {
                "type": "object",
                "description": "Perbarui profil.\n\n============================================================================\n NOMOR HP TIDAK BISA DIUBAH DI SINI\n============================================================================\n Nomor HP adalah identitas login. Mengubahnya menuntut verifikasi OTP ke nomor\n BARU \u2014 kalau tidak, siapa pun yang sempat memegang HP orang lain yang sudah\n login bisa memindahkan akunnya ke nomor sendiri, dan pemilik aslinya\n kehilangan akses beserta seluruh saldonya.\n\n Alurnya ada di endpoint tersendiri dengan purpose `change_phone`.\n============================================================================",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 120
                    },
                    "email": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "email",
                        "description": "Email unik, TAPI mengabaikan baris pengguna ini sendiri. Tanpa `ignore`, pengguna yang menyimpan profilnya tanpa mengubah\nemail akan ditolak dengan \"email sudah dipakai\" \u2014 oleh dirinya\nsendiri. Itu bug yang selalu muncul di form profil dan selalu\nterlihat seperti masalah data.\n\nKolomnya citext, jadi keunikannya sudah tidak peka huruf besar\nkecil di tingkat database.",
                        "maxLength": 255
                    },
                    "gender": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "enum": [
                            "male",
                            "female",
                            "other",
                            null
                        ]
                    },
                    "birth_date": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time",
                        "description": "Tanggal lahir dibatasi ke rentang yang mungkin. Batas 17 tahun bukan pembatasan pemakaian \u2014 anak sekolah memakai\nlayanan ini dengan wajar \u2014 tapi penyaring salah ketik: tanggal\nhari ini atau tahun 1900 hampir selalu berarti pengguna salah\nmemilih di kalender, dan menyimpannya akan membuat laporan\ndemografi tidak berarti."
                    }
                },
                "title": "UpdateProfileRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}